Update screenshot on main page, extract CompetencyNetworkViz into reusable component to use it on main page and while onboarding
